boilerplate/requirements.txt
# Casino Host Agent — Python Dependencies
# Pin major versions for reproducibility. Use >= for minor versions.

# --- LangGraph & LangChain ---
langgraph>=1.0.0,<2.0.0
langgraph-checkpoint>=4.0.0,<5.0.0
langchain-core>=1.0.0,<2.0.0
langchain-community>=0.3.0,<1.0.0
langchain-text-splitters>=0.3.0,<1.0.0

# --- Google AI / GCP ---
langchain-google-genai>=2.0.0,<5.0.0
langchain-google-vertexai>=2.0.0,<3.0.0
google-cloud-firestore>=2.16.0,<3.0.0
google-cloud-aiplatform>=1.60.0,<2.0.0
google-auth>=2.20.0,<3.0.0

# --- API Framework ---
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# --- Observability ---
langsmith>=0.1.0,<1.0.0

# --- Vector DB (local dev) ---
chromadb>=0.5.0,<1.0.0

# --- Utilities ---
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.0.0,<3.0.0
